Minor refactor to make DistPreview.tsx
